+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 20-10-2024
+
+### Added
+
+ - Added logic for DataTransfer API
+ 
+### Changed
+ 
+ - DataTransfer now checks that the Agreement exists and that it's linked to a FINALIZED Contract Negotiation
+
 ## [0.1.1] - 09-10-2024
 
 ### Added